refactored transaction logic

+++ src/main/java/com/chrisimoni/evyntspace/payment/service/TransactionService.java
@@ -1,13 +1,13 @@
 package com.chrisimoni.evyntspace.payment.service;
 
-import com.chrisimoni.evyntspace.payment.enums.CurrencyType;
 import com.chrisimoni.evyntspace.payment.enums.TransactionStatus;
 import com.chrisimoni.evyntspace.payment.model.Transaction;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
 public interface TransactionService {
-    Transaction createTransaction(String paymentIntentId, long amount, String currency, TransactionStatus status);
+    Transaction createTransaction(String paymentIntentId, BigDecimal amount, String currency, TransactionStatus status);
     Optional<Transaction> getTransactionById(UUID transactionId);
 }
